+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Development
 - Fix DWD DMO access again
 - Add lead time argument - one of short, long - for DWD DMO to address two versions of icon
 - Rework dict-like export formats and tests with extensive support for typing
+- Improve radar access
 
 0.65.0 (24.10.2023)
 *******************